+++ Model/SubBill.cs
@@ -10,9 +10,9 @@ namespace Model
     {
         public int SubBillId { get; set; }
         public int BillId { get; set; } // Foreign key
-        public float Price { get; set; }
-        public float Vat { get; set; }
-        public float Tip { get; set; }
+        public decimal Price { get; set; }
+        public decimal Vat { get; set; }
+        public decimal Tip { get; set; }
         public string Feedback { get; set; }
 
         // Navigation property
